### 2023-03-16 #203 @whedon
[PRE REVIEW]: ChooChoo the Checklist tool
**Submitting author:** @lucydot (<a href="http://orcid.org/0000-0002-2992-9871">Lucy Whalley</a>)
**Repository:** https://github.com/lucydot/ChooChoo
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
**Managing EiC:** Jordan Gorzalski
