### 2023-03-17 #203 @lucydot
Hello! A note on reviewers...as this is a domain agnostic tool I think the possible reviewer list is quite broad. Familiarity with Github, Github actions and Python would certainly be useful - and reviewers who uses Github for hosting teaching content might be best as they might have a better appreciation for the project motivation. Looking at the spreadsheet - possible reviewers might be:  lechten gcdeshpande mattforshaw stefania11
