### 2023-03-23 #203 @labarba
@diazrenata Have a look at this submission 👀
Would you be available to edit this? I realize you're more of an R person than Python, but the functionality here is mostly around GitHub... let me know what you think (feel free to ping me on Slack).
