### 2023-07-10 #203 @labarba
Awesome, thank you! FYI, JOSE has been upgraded to the latest more powerful bot: `@editorialbot` — I myself haven't had a chance to try it yet, so here goes...
