### 2023-07-10 #203 @labarba
@ppxasjsm – You had volunteered to contribute a review for JOSE back in May. Are you still able to do it? We've assigned an editor who will look for a second reviewer before starting the Review issue.
