### 2023-07-20 #220 @editorialbot
[REVIEW]: ChooChoo the Checklist tool
**Submitting author:** @lucydot (<a href="http://orcid.org/0000-0002-2992-9871">Lucy Whalley</a>)
**Repository:** https://github.com/lucydot/ChooChoo
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
**Paper kind:** software
