### 2023-08-13 #220 @arfon
@xuanxu – looks like @editorialbot is failing to find the correct checklist when starting these reviews:

```
2023-08-13T07:05:47.372267+00:00 app[web.1]: W, [2023-08-13T07:05:47.372174 #8]  WARN -- : Error calling ReviewerChecklistCommentResponder: GET https://api.github.com/repos/openjournals/jose-reviews/contents/.buffy/templates/reviewer_checklist.md: 404 - Not Found // See: https://docs.github.com/rest/repos/contents#get-repository-content
```

It is looking for `reviewer_checklist.md` rather than a learning module or software checklist. I've gone ahead and edited the checklist for @ggeorg02 but this looks like a bug that needs fixing. Any chance you can take a look please?
