### 2023-09-11 #220 @ggeorg02
I have noticed that the installation document lacks detail. I have identified a few minor issues with the instructions. Despite following all the steps, I was unable to get it to work and received an error message(check at the end of my notes). As a result, I was unable to test the tool's features.

--------------------------
—> Fork an [existing choochoo repository](https://lucydot.github.io/ChooChoo/instructors/setup/existing.md) and use or extend objectives/questions/tutorials from an existing ChooChoo project.

I see it to be something not readable, and a messed up site. Please update it.

—>"Click on Copy repository from template"

This was just “Create repository”

—> “	•	You are advised to create one ChooChoo repository for each class group. This will allow you to track the progress of each individual class rather than all classes combined.
“

Is there anything that an instructor can do if they want to create 20 class groups? I was thinking about group project scenarios where 3 or 4 students in a class form a group. So, should I manually generate for all the student class groups with the current setup?

—> "Enable Issues"

Just wondering if it might makes sense to mention that it is enabled by default(at least, that was the case when I created it)

—> "Enable Workflows"

This was also enabled by default to me

—> 	"Enable Github Pages"
“3	Select gh-pages and /(root) in the drop-down menu below Branch.”

I didn’t see gh-pages branch. Should the user create a branch? Provide details please.

—>  I went with using secrets.GITHUB_TOKEN
—> When running choochoo build checklists
I get error and nothing is printing
[ggeorg02/testachoo] Fat Controller workflow run
