### 2023-09-11 #220 @LittleAprilFool
Same, I have tried multiple approaches to use this tool, including testing it under the author's example repository: https://github.com/lucydot/ChooChoo-template/issues/37, as well as testing it under my own build: https://github.com/LittleAprilFool/choochoo/issues/1. None of the approaches works as the documentation described.
