### 2023-09-13 #220 @lucydot
Hi @LittleAprilFool and @ggeorg02,

Thanks for agreeing to review and test ChooChoo. I suspect that Github has updated its interface and default settings since I submitted this for review 6 months ago which is why some instructions are out of date (also the wrong link, see below, would not have helped!). I will run through the steps myself and update as necessary, and will confirm here once that is done. Thank you for your patience!

This is actually very good timing for me as will be using it again for my own induction week teaching in a fortnight :)

Please note that you ChooChoo does not allow you to run under someone elses repository unless you are added as a student or instructor (this allows repository admin to control who can participate). Of course, you should be able to run under your own so that is what I will look into.

```
—> Fork an [existing choochoo repository](https://lucydot.github.io/ChooChoo/instructors/setup/existing.md) and use or extend objectives/questions/tutorials from an existing ChooChoo project.

I see it to be something not readable, and a messed up site. Please update it.
```

This is a link to the wrong page! Will fix.

Thanks,

Lucy
