### 2023-09-13 dceb6b3
Update setup.md
To reflect recent changes to defaults in github interface
