### 2023-09-13 16fbece
Streamline setup
Make `choochoo-bot` screen name an optional extra - as it is quite a few additional setup steps without much gain. It also requires human intervention which is a potential bottleneck.
