### 2023-09-14 #220 @lucydot
Hi @LittleAprilFool and @ggeorg02,

As suspected the problems were related to update of the Github defaults, including permissions (https://github.blog/changelog/2023-02-02-github-actions-updating-the-default-github_token-permissions-to-read-only/). I've made a few commits to the choochoo-template repository which should fix the problems you were having (and to update the docs) **but** I've not yet had a chance to do a full test of the `choochoo` functionality to make sure that other aspects are working as expected after gh changes, so could I ask you to hold off until next week when I will have a chance to test myself?

I was a little confused why the automatic testing happening at [this repo](https://github.com/choochoo-bot/ChooChoo-testing) didn't flag something up via my github notifications, but then realised that these had been disabled after 60 days of inactivity on that repo (as all workflows to update the fork with any `choochoo `changes and test are scheduled to run daily, so I don't manually change anything there). I'll set myself a reminder to enable the workflows every two months so as to avoid this in future, as a workaround (there may be a neater solution I'm not aware of).

UPDATE: It wasn't the 60 day timeout that caused the problem, it is that the changes to permissions were only applied to new repositories.
