### 2023-10-26 e8926eb
Update setup.md
remove copy existing as an option
