### 2023-10-26 dea7944
Update setup.md
