### 2023-10-26 5bdd6de
Fix re.search for Python 3.11 changes
