### 2023-10-27 7ad831d
Update advanced.md
remove link to pedagogy
