### 2023-10-27 0ae242e
Update setup.md
Update docs so that always specify repository secret with token
