+++ docs/instructors/setup.md
@@ -78,7 +78,7 @@ Commands to `choochoo` trigger a Github Action workflow. By default the action r
 
 === "Personal access token"
 
-1. Generate a personal access token (see docs [here](https://docs.github.com/en/authentication/keeping-your-account-and-data-secure/managing-your-personal-access-tokens)). The only scope required is `public_repo`.
+1. Generate a personal access token [here](https://github.com/settings/tokens/new) (also see docs [here](https://docs.github.com/en/authentication/keeping-your-account-and-data-secure/managing-your-personal-access-tokens)). The only scope required is `public_repo`.
 3. Add the CHOO_ACCESS_TOKEN as a repository secret.
 
 === "Choochoo-bot token"
