### 2023-10-30 #220 @lucydot
Hello @LittleAprilFool, @ggeorg02 and @ttimbers,

Apologies for my testing to take longer than expected - I have been struggling to balance this with the start of a new academic year and lecturing. Thanks for bearing with me.

I have just completed a thorough run-through and located the issues you found.

- as mentioned above, part of the problem was a change in `github_token` permissions which was only applied to new repositories (which was why it wasn't caught in automated testing).
- another problem was related to the `ubuntu-default` used for running github actions. This was recently updated to `ubuntu-22.04` so that the specified python version was no longer supported. I've now updated to `python-3.11` and tested against this.

I have just run through the functionality outlined in the documentation (starting with instructor setup) - https://lucydot.github.io/ChooChoo/instructors/setup/ - and can confirm it ran as expected.

So hopefully you should not have any problems running through now 🤞 . I also used it during induction week with my second year undergraduate class without hiccups (though they did not use full functionality due to lack of time in schedule - they used checklist tool but did not generate their own question sets).

You may note some changes to documentation - I have adjusted some wording to improve readability.

All the best,

Lucy
