### 2025-12-23 #220 @labarba
Hi @LittleAprilFool and @ggeorg02 — Tiffany has stepped down from the JOSE editorial board (she was handling this submission for us). It looks like it was pretty close to the finish line, though, and we needed you to have a new look and consider the author's changes. Could you pop in here and give us your assessment? Many thanks!
