### 2026-03-10 #220 @ggeorg02
@ttimbers / @labarba  I revisited the project and can confirm that the issues I raised during the initial review have been addressed. The documentation has improved significantly and now includes all the relevant details. I was easily able to take on the project, set things up, and get started with the work.
